Commit from Bobby's build
so here is the latest build i have, this is the one from GitHub but with the animation stuff I've worked on today and with Alex plus the quicker movement. Danny has his own build with working Menu and now sounds so we are getting a lot of basics covered and we can use this knowledge for the other areas making things go a lot quicker

for this build arrow keys to move, hold down the Left Shift for running speed, and press Q to attack.
i know its not the best way to be doing things, but @Danny if you wanted to down load this and try adding your menu to it and then put it back up here that would be great.
i also just realized the menu might want to be workable without a mouse as well, I know you have it working for like OnClick but we currently don't use the mouse at all so having the ability to traverse the menu with the arrow keys and like Q to select or something would probably be more along the lines of our final build settings